+++ lib/utilities/readOriginalCW.js
@@ -1,7 +1,7 @@
 /**
- * This script parses Arok's original CW database, which was a set of text files
- * he wrote before beginning to use Toolbox. These files are located in the ALTLab
- * repo under `crk/dicts/CW_original`.
+ * This script parses Arok's original CW database, which was a set of Word Perfect
+ * text files he wrote before beginning to use Toolbox. These files are located
+ * in the ALTLab repo under `crk/dicts/CW_original`.
  */
 
 import { promises as fsPromises } from 'fs';
