+++ package.json
@@ -22,16 +22,16 @@
   },
   "type": "module",
   "scripts": {
-    "build": "node bin/buildDatabase.js",
-    "clear": "node bin/clearDatabase.js",
-    "test": "mocha lib/**/*.test.js test/*.test.js",
+    "build":      "node bin/buildDatabase.js",
+    "clear":      "node bin/clearDatabase.js",
+    "test":       "mocha lib/**/*.test.js test/*.test.js",
     "test:build": "mocha lib/testBuild.js"
   },
   "bin": {
-    "convert-cw": "./bin/convert-CW.js",
-    "convert-md": "./bin/convert-MD.js",
-    "import-cw": "./bin/import-CW.js",
-    "import-md": "./bin/import-MD.js",
+    "convert-cw":        "./bin/convert-CW.js",
+    "convert-md":        "./bin/convert-MD.js",
+    "import-cw":         "./bin/import-CW.js",
+    "import-md":         "./bin/import-MD.js",
     "import-cw-sources": "./bin/import-CW-sources.js"
   },
   "devDependencies": {
