+++ test/CW.test.db
@@ -1,61 +1,5 @@
 \_sh v1.0 9999 Plains Cree
 \_DateStampHasFourDigitYear
 
-\test assigns homograph numbers
-\sro  kâkito
-
-\test assigns homograph numbers
-\sro  kâkito
-
-\test combines multiple dialect codes
-\dl   Cree: pC
-\dl   Cree: sC
-
-\test converts dialects to Glottocodes
-\dl   Cree: npC
-
-\test copies the glosses as an Array
-\gl   star
-\gl   little
-
-\test copies the original definition field verbatim
-\def  North Star, Polaris; [lit: "star that does not move"]
-
-\test copies the POS field verbatim
-\ps   VAI-1
-
-\test copies the SRO field verbatim
-\sro  acicipaýihow
-
-\test copies the Syllabics field verbatim
-\syl  ᐊᒑᐦᑯᐢ  ᐁᑳ  ᑳ ᐋᐦᒌᐟ
-
-\test extracts compare relations: [cf. XXX]
-\def  not [cf. nama]
-
-\test extracts cross-references: [see XXX]
-\def  aside, off to one side [see yîkatê-]
-
-\test extracts extended cross-references as notes: [see XXX …]
-\def  greatly, extremely, overwhelmingly, tremendously, to an extraordinary extent; with startled surprise [see kakwâýaki- for further derivatives]
-
-\test extracts general notes
-\def  if only it were so, if only; of course not, not any, not even; [i.e. foregone conclusion that something is not the case]
-
-\test extracts Latin terms
-\def  black ash [Lt. Fraximus nigra]
-
-\test extracts multiple notes from a parenthetical
-\def  dock [Lt. Rumex sp.; literally: "yellow-root"]
-
-\test extracts multiple parentheticals
-\def  young macôhow, Bad Owl's son [male name]; [plural:] Bad Owl's family, members of Bad Owl's family
-
-\test extracts unbracketed literal definitions
-\def  Christmas moon; December; literally: "God-day moon, Christmas moon"
-
-\test extracts usage notes
-\def  insistently; [in negative clauses:] (not) necessarily
-
-\test separates definitions by semicolons
-\def  it is a star; s/he is a star (e.g. in movies, sports, music, etc.)
+\test test name
+\sro  test-data
